requirements.txt
agora-token-builder==1.0.0
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
django-browser-reload==1.12.1
django-social-share==2.3.0
django-tailwind==3.6.0
Django-Verify-Email==2.0.3
gunicorn==21.2.0
idna==3.4
joblib==1.3.2
numpy==1.26.2
packaging==23.2
Pillow==10.1.0
reportlab==4.0.7
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
sqlparse==0.4.4
threadpoolctl==3.2.0
tzdata==2023.3
urllib3==2.1.0
